requirements.txt
streamlit
pandas
camelot-py
PyPDF2  
openpyxl  
